refactor: move math things to the separate package
