+++ slider/src/main/java/io/monstarlab/mosaic/slider/distribution/CheckPointsValueDistribution.kt
@@ -1,6 +1,9 @@
 package io.monstarlab.mosaic.slider.distribution
 
-import io.monstarlab.mosaic.slider.valueToFraction
+import io.monstarlab.mosaic.slider.math.LinearEquation
+import io.monstarlab.mosaic.slider.math.Point
+import io.monstarlab.mosaic.slider.math.RangedLinearEquation
+import io.monstarlab.mosaic.slider.math.valueToFraction
 
 public class CheckPointsValueDistribution(
     valuesMap: List<Pair<Float, Float>>,
@@ -16,31 +19,29 @@ public class CheckPointsValueDistribution(
 
         val offsetRange = valuesMap.minOf { it.first }..valuesMap.maxOf { it.first }
         val valueRange = valuesMap.minOf { it.second }..valuesMap.maxOf { it.second }
-        val zipped = valuesMap.sortedBy { it.first }
+
+        equations = valuesMap.sortedBy { it.first }
             .zipWithNext()
-        // make sure all values are increasing, otherwise throw an exception
-        zipped.firstOrNull { it.first.second >= it.second.second }
-            ?.let {
-                throw DecreasingValueException(it.first)
+            .checkIncreasingValues() // check if values are always increasing
+            .map {
+                val x1Fraction = it.first.first.valueToFraction(offsetRange)
+                val x2Fraction = it.second.first.valueToFraction(offsetRange)
+                val y1Fraction = it.first.second.valueToFraction(valueRange)
+                val y2Fraction = it.second.second.valueToFraction(valueRange)
+                val equation = LinearEquation.fromTwoPoints(
+                    x1 = x1Fraction,
+                    x2 = x2Fraction,
+                    y1 = y1Fraction,
+                    y2 = y2Fraction,
+                )
+                RangedLinearEquation(
+                    equation = equation,
+                    offsetRange = x1Fraction..x2Fraction,
+                    valueRange = y1Fraction..y2Fraction,
+                )
             }
-        equations = zipped.map {
-            val x1Fraction = it.first.first.valueToFraction(offsetRange)
-            val x2Fraction = it.second.first.valueToFraction(offsetRange)
-            val y1Fraction = it.first.second.valueToFraction(valueRange)
-            val y2Fraction = it.second.second.valueToFraction(valueRange)
-            val equation = LinearEquation.fromTwoPoints(
-                x1 = x1Fraction,
-                x2 = x2Fraction,
-                y1 = y1Fraction,
-                y2 = y2Fraction,
-            )
-            RangedLinearEquation(
-                equation = equation,
-                offsetRange = x1Fraction..x2Fraction,
-                valueRange = y1Fraction..y2Fraction,
-            )
-        }
     }
+
 
     override fun interpolate(value: Float): Float {
         val equation = equations.firstOrNull { it.offsetRange.contains(value) }?.equation
@@ -54,10 +55,17 @@ public class CheckPointsValueDistribution(
         return equation.offsetFromValue(value)
     }
 
-    public class DecreasingValueException(progressValuePair: Pair<Float, Float>) :
+    private fun List<Pair<Point, Point>>.checkIncreasingValues(): List<Pair<Point, Point>> {
+        find { it.first.second >= it.second.second }?.let {
+                throw DecreasingValueException(it.first)
+        }
+        return this
+    }
+
+    public class DecreasingValueException(progressValuePair: Point) :
         IllegalStateException(
             "Values must be always increasing with increasing progress," +
-                " item at progress ${progressValuePair.first}  with value " +
-                "${progressValuePair.second} is breaking this rule ",
+                    " item at progress ${progressValuePair.first}  with value " +
+                    "${progressValuePair.second} is breaking this rule ",
         )
 }
