+++ slider/src/main/java/io/monstarlab/mosaic/slider/math/MathUtils.kt
@@ -1,4 +1,4 @@
-package io.monstarlab.mosaic.slider
+package io.monstarlab.mosaic.slider.math
 
 import androidx.compose.ui.util.lerp
 
@@ -15,7 +15,7 @@ internal fun Float.valueToFraction(range: ClosedFloatingPointRange<Float>) =
     valueToFraction(range.start, range.endInclusive)
 
 internal fun Float.fractionToValue(rangeStart: Float, rangeEnd: Float): Float =
-    scale(0f, 1f, this.coerceIn(0f, 1f), rangeStart, rangeEnd)
+    scale(0f, 1f, coerceIn(0f, 1f), rangeStart, rangeEnd)
 
 internal fun Float.fractionToValue(range: ClosedFloatingPointRange<Float>): Float =
-    scale(0f, 1f, this, range.start, range.endInclusive)
+    fractionToValue(range.start, range.endInclusive)
