+++ slider/src/main/java/io/monstarlab/mosaic/slider/math/Point.kt
@@ -1,0 +1,3 @@
+package io.monstarlab.mosaic.slider.math
+
+internal typealias Point = Pair<Float, Float>
